libs/auth/pyproject.toml
[project]
name = "amortsched-auth"
version = "0.1.0"
description = "Authentication adapters (JWT)"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "amortsched-core",
    "python-jose[cryptography]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/amortsched"]
